+++ src/helpers/noop-utils.ts
@@ -104,7 +104,7 @@ export const noopPromiseReject = (): Promise<never> => Promise.reject();
 export const noopPromiseResolve = (
     responseBody = '{}',
     responseUrl = '',
-    responseType = 'default',
+    responseType = 'basic',
 ): Promise<Response> | undefined => {
     if (typeof Response === 'undefined') {
         return;
@@ -115,21 +115,27 @@ export const noopPromiseResolve = (
         statusText: 'OK',
     });
 
-    // Mock response' url & type to avoid adb checks
-    // https://github.com/AdguardTeam/Scriptlets/issues/216
-    Object.defineProperties(response, {
-        url: { value: responseUrl },
-        type: { value: responseType },
-    });
-
-    // In the case if responseType is opaque
-    // mock response' body, status & statusText to avoid adb checks
-    // https://github.com/AdguardTeam/Scriptlets/issues/364
+    /**
+     * If responseType is opaque, then by default
+     * body is null, status is 0, statusText and url are empty strings,
+     * as per https://fetch.spec.whatwg.org/#concept-filtered-response-opaque
+     * It's required to mock these properties to avoid adblock detection
+     * https://github.com/AdguardTeam/Scriptlets/issues/364
+     */
     if (responseType === 'opaque') {
         Object.defineProperties(response, {
             body: { value: null },
             status: { value: 0 },
             statusText: { value: '' },
+            url: { value: '' },
+            type: { value: responseType },
+        });
+    } else {
+        // Mock response' url & type to avoid adb checks
+        // https://github.com/AdguardTeam/Scriptlets/issues/216
+        Object.defineProperties(response, {
+            url: { value: responseUrl },
+            type: { value: responseType },
         });
     }